api_gateway/requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
grpcio==1.70.0
grpcio-tools==1.70.0
python-dotenv==0.19.1
email-validator==2.2.0
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-cov==4.1.0